added atomic design

+++ src/components/molecules/answersList/answersList.js
@@ -1,28 +1,33 @@
 import React from "react";
 import styles from "./answersList.module.css";
 import moment from "moment";
-import DeleteButton from "../deleteButton/deleteButton";
-import LikeDislike from "../likeDislike/likeDislike";
+import DeleteButton from "../../atoms/deleteButton/deleteButton";
+import LikeDislike from "../../atoms/likeDislike/likeDislike";
 
 const AnswersList = (allAnswers) => {
   const answers = allAnswers.allAnswers.questionAnswers;
-  const token = window.localStorage.token;
+
   const handleDelete = (answer) => {
-    const id = {
-      id: answer._id,
-    };
-    fetch(`http://localhost:8000/answer/${answer._id}`, {
-      method: "DELETE",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-        Authorization: token,
-      },
-      body: JSON.stringify(id),
-    }).then((res) => {
-      return res.json();
-    });
-    window.location.reload(false);
+    const token = window.localStorage.token;
+    if (token) {
+      const id = {
+        id: answer._id,
+      };
+      fetch(`http://localhost:8000/answer/${answer._id}`, {
+        method: "DELETE",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+        body: JSON.stringify(id),
+      }).then((res) => {
+        return res.json();
+      });
+      window.location.reload(false);
+    } else {
+      alert("Please login or sign up to ask a question!");
+    }
   };
   return (
     <div className={styles.main}>
