+++ src/components/molecules/question/question.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "./question.module.css";
-import DeleteButton from "../deleteButton/deleteButton";
+import DeleteButton from "../../atoms/deleteButton/deleteButton";
 import { useNavigate } from "react-router-dom";
 import moment from "moment";
 
@@ -11,30 +11,33 @@ const Question = ({ question }) => {
   };
   const date = moment(question.dateCreated).fromNow();
 
-  const token = window.localStorage.token;
   const handleDelete = () => {
-    const id = {
-      id: question._id,
-    };
-    fetch(`http://localhost:8000/question/${question.id}`, {
-      method: "DELETE",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-        Authorization: token,
-      },
-      body: JSON.stringify(id),
-    }).then((res) => {
-      return res.json();
-    });
-    window.location.reload(false);
+    const token = window.localStorage.token;
+    if (token) {
+      const id = {
+        id: question._id,
+      };
+      fetch(`http://localhost:8000/question/${question.id}`, {
+        method: "DELETE",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+        body: JSON.stringify(id),
+      }).then((res) => {
+        return res.json();
+      });
+      window.location.reload(false);
+    } else {
+      alert("Please login or sign up to ask a question!");
+    }
   };
   return (
     <div className={styles.main} onClick={() => onClickComponent()}>
       <h1>{question.title}</h1>
       <div className={styles.rightContainer}>
         <h5>{date}</h5>
-        {/* <h6>{answered}</h6> */}
         <DeleteButton text="X" onClick={() => handleDelete()} />
       </div>
     </div>
